requirements.txt
fastapi
uvicorn
groq
playwright
moviepy
edge-tts
pandas
jinja2
python-dotenv
aiohttp
aiofiles
jinja2
python-multipart
imageio[pyav]
